requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
autograd==1.5
autoray==0.6.12
bidict==0.23.1
boto3==1.34.143
botocore==1.34.143
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
configparser==7.0.0
covalent==0.232.0.post1
covalent-api-sdk==1.0.2
covalent-python-api==0.0.4
dask==2024.7.0
dateparser==1.2.0
Deprecated==1.2.14
distributed==2024.7.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
furl==2.1.3
future==1.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
lmdb==1.5.1
lmdbm==0.0.6
locket==1.0.0
Mako==1.3.5
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpire==2.10.2
msgpack==1.0.8
multidict==6.0.5
mysql-connector==2.2.9
mysqlclient==2.2.4
natsort==8.4.0
networkx==3.2.1
numpy==1.23.5
orderedmultidict==1.0.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
partd==1.4.2
PennyLane==0.32.0
PennyLane-Lightning==0.33.1
pkgconfig==1.5.5
pluggy==1.5.0
psutil==6.0.0
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyMySQL==1.1.1
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-env==1.1.3
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rich==13.3.5
rustworkx==0.15.1
s3transfer==0.10.2
scipy==1.13.1
semantic-version==2.10.0
shellingham==1.5.4
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
tblib==3.0.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.19
uvicorn==0.18.3
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zict==3.0.0
zipp==3.19.2
